move problem validation classes to problem module
